@@ -1,31 +1,32 @@
 {
-  "name": "example",
-  "private": true,
-  "version": "1.0.0",
-  "license": "MIT",
-  "scripts": {
-    "start": "cross-env NODE_ENV=production vite --port=3000",
-    "build": "vite build"
-  },
   "dependencies": {
-    "@tanstack/react-query": "^4.24.6",
-    "@tanstack/react-query-devtools": "^4.24.6",
+    "@tanstack/react-query": "^5.62.3",
+    "@tanstack/react-query-devtools": "^5.62.3",
     "cross-env": "^7.0.3",
-    "msw": "^1.0.1",
+    "msw": "^2.6.8",
     "react-app-polyfill": "^3.0.0",
-    "react-query-auth": "2.1.1"
+    "react-query-auth": "*"
   },
   "devDependencies": {
-    "@types/node": "^18.14.0",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "typescript": "^4",
+    "@types/node": "^22.10.1",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "typescript": "^5",
     "vite": "latest",
     "vite-preset-react": "latest"
   },
+  "license": "MIT",
   "msw": {
     "workerDirectory": "public"
-  }
+  },
+  "name": "example",
+  "private": true,
+  "scripts": {
+    "build": "vite build",
+    "dev": "vite",
+    "start": "cross-env NODE_ENV=production vite --port=3000"
+  },
+  "version": "1.0.0"
 }
