@@ -1,71 +1,71 @@
 {
-  "name": "react-query-auth",
-  "version": "2.4.1",
-  "author": "alan2207",
-  "license": "MIT",
-  "repository": "alan2207/react-query-auth",
-  "main": "./dist/index.mjs",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "exports": {
-    "require": "./dist/index.js",
-    "import": "./dist/index.mjs",
-    "types": "./dist/index.d.ts"
-  },
-  "engines": {
-    "node": ">=16"
-  },
-  "scripts": {
-    "dev": "tsup src/index.tsx --watch",
-    "build": "tsup src/index.tsx",
-    "test": "jest",
-    "lint": "eslint --fix ./src",
-    "size": "size-limit"
-  },
-  "peerDependencies": {
-    "@tanstack/react-query": ">=4.24.6",
-    "react": ">=16.8.0"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "npm run lint"
-    }
-  },
-  "size-limit": [
-    {
-      "path": "dist/index.mjs",
-      "limit": "10 KB"
-    },
-    {
-      "path": "dist/index.js",
-      "limit": "10 KB"
-    }
-  ],
-  "devDependencies": {
-    "@size-limit/preset-small-lib": "^8.2.4",
-    "@tanstack/react-query": "^5.7.2",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/react-hooks": "^8.0.1",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.52.0",
-    "@typescript-eslint/parser": "^5.52.0",
-    "eslint": "^8.34.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-react": "^7.32.2",
-    "husky": "^8.0.3",
-    "jest": "^29.4.3",
-    "jest-environment-jsdom": "^29.4.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "size-limit": "^8.2.4",
-    "ts-jest": "^29.0.5",
-    "tsup": "^6.6.3",
-    "typescript": "^4.9.5"
-  }
+	"name": "react-query-auth",
+	"version": "2.4.1",
+	"author": "alan2207",
+	"license": "MIT",
+	"repository": "alan2207/react-query-auth",
+	"main": "./dist/index.js",
+	"module": "./dist/index.mjs",
+	"types": "./dist/index.d.mts",
+	"exports": {
+		"./package.json": "./package.json",
+		".": {
+			"import": {
+				"types": "./dist/index.d.mts",
+				"import": "./dist/index.mjs",
+				"default": "./dist/index.mjs"
+			},
+			"require": {
+				"types": "./dist/index.d.ts",
+				"import": "./dist/index.js",
+				"require": "./dist/index.js"
+			}
+		}
+	},
+	"files": ["dist"],
+	"engines": {
+		"node": ">=20.0.0"
+	},
+	"scripts": {
+		"test": "vitest run",
+		"test:cov": "vitest run --coverage",
+		"postbuild": "npm run check:exports",
+		"build": "tsup src/index.tsx --config tsup.config.ts",
+		"example:dev": "npm run dev -w examples/vite",
+		"build:dev": "tsup src/index.tsx --config tsup.dev.config.ts",
+		"build:dev:watch": "npm run build:dev -- --watch",
+		"build:dev:cjs:watch": "npm run build:dev -- --watch",
+		"dev": "npm-run-all -s build:dev -p example:dev build:dev:watch",
+		"prepublishOnly": "npm run build",
+		"check": "biome check .",
+		"check:fix": "biome check --fix .",
+		"typecheck": "tsc",
+		"validate": "npm run check && npm run tsc && npm run test",
+		"check:exports": "attw --pack ."
+	},
+	"peerDependencies": {
+		"@tanstack/react-query": ">=4.24.6",
+		"react": ">=16.8.0"
+	},
+	"devDependencies": {
+		"@arethetypeswrong/cli": "^0.17.1",
+		"@biomejs/biome": "^1.9.4",
+		"@tanstack/react-query": "^5.62.3",
+		"@testing-library/jest-dom": "^6.6.3",
+		"@testing-library/react": "^16.1.0",
+		"@types/node": "^22.10.1",
+		"@types/react": "^18.0.28",
+		"@types/react-dom": "^18.0.11",
+		"@vitest/coverage-v8": "^2.1.8",
+		"happy-dom": "^15.11.7",
+		"jsdom": "^25.0.1",
+		"lefthook": "^1.9.0",
+		"npm-run-all": "^4.1.5",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"tsup": "^8.3.5",
+		"typescript": "^5.7.2",
+		"vitest": "^2.1.8"
+	},
+	"workspaces": [".", "examples/*"]
 }
