@@ -1,55 +1,54 @@
-import { configureAuth } from 'react-query-auth';
+import { configureAuth } from "react-query-auth"
 import {
-  getUserProfile,
-  registerWithEmailAndPassword,
-  loginWithEmailAndPassword,
-  AuthResponse,
-  logout,
-} from './api';
-import { storage } from './utils';
+	type AuthResponse,
+	getUserProfile,
+	loginWithEmailAndPassword,
+	logout,
+	registerWithEmailAndPassword,
+} from "./api"
+import { storage } from "./utils"
 
 export type LoginCredentials = {
-  email: string;
-  password: string;
-};
+	email: string
+	password: string
+}
 
 export type RegisterCredentials = {
-  email: string;
-  name: string;
-  password: string;
-};
+	email: string
+	name: string
+	password: string
+}
 
 async function handleUserResponse(data: AuthResponse) {
-  const { jwt, user } = data;
-  storage.setToken(jwt);
-  return user;
+	const { jwt, user } = data
+	storage.setToken(jwt)
+	return user
 }
 
 async function userFn() {
-  const { user } = await getUserProfile();
-  return user ?? null;
+	const { user } = await getUserProfile()
+	return user ?? null
 }
 
 async function loginFn(data: LoginCredentials) {
-  const response = await loginWithEmailAndPassword(data);
-  const user = await handleUserResponse(response);
-  return user;
+	const response = await loginWithEmailAndPassword(data)
+	const user = await handleUserResponse(response)
+	return user
 }
 
 async function registerFn(data: RegisterCredentials) {
-  const response = await registerWithEmailAndPassword(data);
-  const user = await handleUserResponse(response);
-  return user;
+	const response = await registerWithEmailAndPassword(data)
+	const user = await handleUserResponse(response)
+	return user
 }
 
 async function logoutFn() {
-  await logout();
+	await logout()
 }
 
-export const { useUser, useLogin, useRegister, useLogout, AuthLoader } =
-  configureAuth({
-    userFn,
-    loginFn,
-    registerFn,
-    logoutFn,
-  });
+export const { useUser, useLogin, useRegister, useLogout, AuthLoader } = configureAuth({
+	userFn,
+	loginFn,
+	registerFn,
+	logoutFn,
+})
